+++ src/com/garbagemule/MobArena/health/HealthStrategyStandard.java
@@ -5,7 +5,7 @@ import org.bukkit.entity.Player;
 public class HealthStrategyStandard implements HealthStrategy
 {
     @Override
-    public void setHealth(Player p, int health) {
+    public void setHealth(Player p, double health) {
         p.setHealth(health);
     }
 }
